Move `setAriaDescribedBy` to `helpers`
