+++ packages/components/addon/components/hds/form/field/index.js
@@ -8,7 +8,7 @@ import { tracked } from '@glimmer/tracking';
 import { assert } from '@ember/debug';
 import { action } from '@ember/object';
 import { getElementId } from '@hashicorp/design-system-components/helpers/hds-get-element-id';
-import { setAriaDescribedBy } from '../../../../helpers/hds-set-aria-described-by';
+import { setAriaDescribedBy } from '@hashicorp/design-system-components/helpers/hds-set-aria-described-by';
 import { schedule } from '@ember/runloop';
 
 export const LAYOUT_TYPES = ['vertical', 'flag'];
