+++ packages/components/addon/components/hds/form/radio-card/index.js
@@ -7,7 +7,7 @@ import Component from '@glimmer/component';
 import { tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
 import { assert } from '@ember/debug';
-import { setAriaDescribedBy } from '../utils/setAriaDescribedBy';
+import { setAriaDescribedBy } from '@hashicorp/design-system-components/helpers/hds-set-aria-described-by';
 import { schedule } from '@ember/runloop';
 
 export const DEFAULT_CONTROL_POSITION = 'bottom';
